TweetsContainer has a list of tweets
This means TweetsContainer is no longer a Box, it is now a List.

Tweet is not a box, but rather just an object to keep record of an
entire tweet.  Eventually it'd have methods of course.

Screen is simpler.  Now it's just a factory for the blessed.screen
object, it no longer adds its own methods.